Implement patient eligibility verification functionality: - Add EligibilityController with check endpoint - Create EligibilityService interface and implementation - Add eligibility request/response DTOs - Implement logic for checking patient insurance status - Add service coverage verification - Support dependent eligibility checking

+++ src/main/java/com/medco/HealthConnectProvider/entity/services/Servicelist.java
@@ -48,7 +48,7 @@ public class Servicelist extends Audit implements Serializable {
     @Enumerated(EnumType.STRING)
     private Status status;
 
-    private Double price;
+    private BigDecimal price;
 
     // Many-to-One relationship with Provider
     @ManyToOne(fetch = FetchType.LAZY)
